Renamed MainPage route to SynopticPage
